Fix test_run_datastore_analysis test.
This fixes the SSA on datastores by navigating to the datastores from
provider and then starting the test.

Also
Resolve circular imports

Fixes:

  File "/cfme/cfme_tests/cfme/utils/providers.py", line 16, in <module>
    from cfme.common.provider import all_types
ImportError: Error importing plugin "cfme.markers.env": cannot import name 'all_types' from 'cfme.common.provider' (/cfme/cfme_tests/cfme/common/provider.py)
