+++ cfme/common/provider.py
@@ -17,6 +17,7 @@ from cfme.base.credential import SSHCredential
 from cfme.base.credential import TokenCredential
 from cfme.common import CustomButtonEventsMixin
 from cfme.common import Taggable
+from cfme.common.datastore_views import ProviderAllDatastoresView
 from cfme.exceptions import AddProviderError
 from cfme.exceptions import HostStatsNotContains
 from cfme.exceptions import ProviderHasNoKey
@@ -26,6 +27,7 @@ from cfme.modeling.base import BaseEntity
 from cfme.utils import conf
 from cfme.utils import ParamClassName
 from cfme.utils.appliance import Navigatable
+from cfme.utils.appliance.implementations.ui import CFMENavigateStep
 from cfme.utils.appliance.implementations.ui import navigate_to
 from cfme.utils.appliance.implementations.ui import navigator
 from cfme.utils.log import logger
@@ -1374,3 +1376,14 @@ class DefaultEndpointForm(View):
     change_password = Text(locator='.//a[normalize-space(.)="Change stored password"]')
 
     validate = Button('Validate')
+
+
+@navigator.register(BaseProvider, 'DatastoresOfProvider')
+class DatastoresOfProvider(CFMENavigateStep):
+    VIEW = ProviderAllDatastoresView
+
+    def prerequisite(self):
+        return navigate_to(self.obj, 'Details')
+
+    def step(self, *args, **kwargs):
+        self.prerequisite_view.entities.summary('Relationships').click_at('Datastores')
